python/requirements.txt
python-dotenv

fastapi
uvicorn[standard]
pydantic

openai

langchain-community
langchain-openai
langchain-text-splitters

faiss-cpu
